newrequirements.txt
bottle==0.12.25
bottle-websocket==0.2.9
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.0.7
cycler==0.11.0
Eel==0.16.0
fonttools==4.39.3
future==0.18.3
gevent==23.9.1
gevent-websocket==0.10.1
greenlet==3.0.1
gTTS==2.5.1
idna==3.6
kiwisolver==1.4.4
matplotlib==3.7.1
MouseInfo==0.1.3
numpy==1.24.3
packaging==23.1
Pillow==9.5.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyobjc-core==10.0
pyobjc-framework-Cocoa==10.0
pyobjc-framework-Quartz==10.0
pyparsing==3.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
pyserial==3.5
python-dateutil==2.8.2
python-vlc==3.0.20123
pytweening==1.0.7
requests==2.31.0
rubicon-objc==0.4.7
six==1.16.0
urllib3==2.2.0
whichcraft==0.6.1
zope.event==5.0
zope.interface==6.1
